divide the dashboard_canvas_content file
