+++ x-pack/platform/plugins/shared/dashboard_agent/public/attachment_types/dashboard_canvas_content.tsx
@@ -6,259 +6,19 @@
  */
 
 import React, { useCallback, useMemo, useState } from 'react';
-import { EuiButton, EuiFlexGroup, EuiFlexItem, useEuiTheme } from '@elastic/eui';
+import { EuiButton, EuiFlexGroup, EuiFlexItem } from '@elastic/eui';
 import { css } from '@emotion/react';
 import type { AttachmentRenderProps } from '@kbn/agent-builder-browser/attachments';
-import type { DashboardAttachmentData, AttachmentPanel } from '@kbn/dashboard-agent-common';
-import { isGenericAttachmentPanel, isLensAttachmentPanel } from '@kbn/dashboard-agent-common';
+import type { DashboardAttachmentData } from '@kbn/dashboard-agent-common';
 import type { DashboardAttachment } from '@kbn/dashboard-agent-common/types';
 import type { DashboardState } from '@kbn/dashboard-plugin/common';
 import { DashboardRenderer } from '@kbn/dashboard-plugin/public';
 import type { DashboardApi, DashboardCreationOptions } from '@kbn/dashboard-plugin/public';
-import type { DashboardPanel } from '@kbn/dashboard-plugin/server';
-import { MARKDOWN_EMBEDDABLE_TYPE } from '@kbn/dashboard-markdown/public';
 import { i18n } from '@kbn/i18n';
-import {
-  LensConfigBuilder,
-  type LensApiSchemaType,
-  type LensAttributes,
-} from '@kbn/lens-embeddable-utils/config_builder';
-import { isLensLegacyAttributes } from '@kbn/lens-embeddable-utils/config_builder/utils';
-import type { LensSerializedAPIConfig } from '@kbn/lens-common-2';
-import { LENS_EMBEDDABLE_TYPE } from '@kbn/lens-plugin/public';
+import { useMemoCss } from '@kbn/css-utils/public/use_memo_css';
+import { normalizePanels } from './panel_grid_layout';
 
-const DASHBOARD_GRID_COLUMN_COUNT = 48;
-const DEFAULT_PANEL_HEIGHT = 9;
-const SMALL_PANEL_WIDTH = 12;
-const LARGE_PANEL_WIDTH = 24;
-const MARKDOWN_PANEL_WIDTH = 48;
-const MARKDOWN_MIN_HEIGHT = 6;
-const MARKDOWN_MAX_HEIGHT = 9;
-const SMALL_CHART_TYPES = new Set(['metric', 'legacy_metric', 'gauge']);
-
-export const PANEL_LAYOUT = {
-  defaultPanelHeight: DEFAULT_PANEL_HEIGHT,
-  smallPanelWidth: SMALL_PANEL_WIDTH,
-  largePanelWidth: LARGE_PANEL_WIDTH,
-  markdownPanelWidth: MARKDOWN_PANEL_WIDTH,
-  markdownMinHeight: MARKDOWN_MIN_HEIGHT,
-  markdownMaxHeight: MARKDOWN_MAX_HEIGHT,
-  smallChartTypes: SMALL_CHART_TYPES,
-  dashboardGridColumnCount: DASHBOARD_GRID_COLUMN_COUNT,
-};
-
-export const getLensPanelWidthFromAttributes = (lensAttributes: LensAttributes): number => {
-  const visType = lensAttributes.visualizationType;
-  const isSmallChart =
-    visType === 'lnsMetric' || visType === 'lnsLegacyMetric' || visType === 'lnsGauge';
-  return isSmallChart ? SMALL_PANEL_WIDTH : LARGE_PANEL_WIDTH;
-};
-
-export const getPanelWidth = (chartType: string, layout = PANEL_LAYOUT): number => {
-  return layout.smallChartTypes.has(chartType) ? layout.smallPanelWidth : layout.largePanelWidth;
-};
-
-export const calculateMarkdownPanelHeight = (content: string, layout = PANEL_LAYOUT): number => {
-  const lineCount = content.split('\n').length;
-  const estimatedHeight = lineCount + 2;
-  return Math.max(layout.markdownMinHeight, Math.min(layout.markdownMaxHeight, estimatedHeight));
-};
-
-export const buildLensPanelFromApi = (
-  config: LensApiSchemaType,
-  grid: DashboardPanel['grid'],
-  uid?: string
-): DashboardPanel => {
-  const lensAttributes: LensAttributes = new LensConfigBuilder().fromAPIFormat(config);
-
-  const lensConfig: LensSerializedAPIConfig = {
-    title:
-      lensAttributes.title ??
-      config.title ??
-      i18n.translate('xpack.dashboardAgent.attachments.dashboard.generatedPanelTitle', {
-        defaultMessage: 'Generated panel',
-      }),
-    attributes: lensAttributes,
-  };
-
-  return {
-    type: 'lens',
-    grid,
-    config: lensConfig,
-    uid,
-  };
-};
-
-export const isLensEmbeddableType = (
-  embeddableType: string,
-  rawConfig: unknown
-): rawConfig is LensAttributes => {
-  return embeddableType === LENS_EMBEDDABLE_TYPE && isLensLegacyAttributes(rawConfig);
-};
-
-export interface BuildPanelFromRawConfigOptions {
-  embeddableType: string;
-  rawConfig: Record<string, unknown>;
-  title: string | undefined;
-  position: { currentX: number; currentY: number };
-  layout?: typeof PANEL_LAYOUT;
-  uid?: string;
-  getLensPanelWidth?: (lensAttributes: LensAttributes) => number;
-  /** When set (e.g. from agent-specified layout), use these dimensions instead of layout defaults. */
-  preferredGrid?: { w: number; h: number };
-}
-
-export const buildPanelFromRawConfig = ({
-  layout = PANEL_LAYOUT,
-  embeddableType,
-  rawConfig,
-  title,
-  position,
-  uid,
-  getLensPanelWidth,
-  preferredGrid,
-}: BuildPanelFromRawConfigOptions): DashboardPanel | null => {
-  const { currentX, currentY } = position;
-
-  if (isLensEmbeddableType(embeddableType, rawConfig)) {
-    const lensAttributes = rawConfig;
-    const width =
-      preferredGrid?.w ??
-      (getLensPanelWidth ? getLensPanelWidth(lensAttributes) : layout.largePanelWidth);
-    const height = preferredGrid?.h ?? layout.defaultPanelHeight;
-
-    let x = currentX;
-    let y = currentY;
-    if (x + width > layout.dashboardGridColumnCount) {
-      x = 0;
-      y += layout.defaultPanelHeight;
-    }
-
-    const lensConfig: LensSerializedAPIConfig = {
-      title: title ?? lensAttributes.title ?? 'Panel',
-      attributes: lensAttributes,
-    };
-
-    return {
-      type: 'lens',
-      grid: { x, y, w: width, h: height },
-      config: lensConfig,
-      uid,
-    };
-  }
-
-  if (embeddableType === MARKDOWN_EMBEDDABLE_TYPE) {
-    const content = (rawConfig as { content?: string }).content ?? '';
-    const height = preferredGrid?.h ?? calculateMarkdownPanelHeight(content, layout);
-    const width = preferredGrid?.w ?? layout.markdownPanelWidth;
-
-    let x = currentX;
-    let y = currentY;
-    if (x + width > layout.dashboardGridColumnCount) {
-      x = 0;
-      y += layout.defaultPanelHeight;
-    }
-
-    return {
-      type: MARKDOWN_EMBEDDABLE_TYPE,
-      grid: { x, y, w: width, h: height },
-      config: { content },
-      uid,
-    };
-  }
-
-  const width = preferredGrid?.w ?? layout.largePanelWidth;
-  const panelHeight = preferredGrid?.h ?? layout.defaultPanelHeight;
-  let x = currentX;
-  let y = currentY;
-  if (x + width > layout.dashboardGridColumnCount) {
-    x = 0;
-    y += layout.defaultPanelHeight;
-  }
-
-  return {
-    type: embeddableType,
-    grid: { x, y, w: width, h: panelHeight },
-    config: rawConfig,
-    uid,
-  };
-};
-
-export const normalizePanels = (
-  panels: AttachmentPanel[],
-  yOffset: number = 0,
-  layout: typeof PANEL_LAYOUT = PANEL_LAYOUT,
-  includePanelIdAsUid: boolean = true,
-  getLensPanelWidth: (lensAttributes: LensAttributes) => number = getLensPanelWidthFromAttributes
-): DashboardPanel[] => {
-  const normalizedLayout = layout ?? PANEL_LAYOUT;
-  const panelList = panels ?? [];
-  const dashboardPanels: DashboardPanel[] = [];
-  let currentX = 0;
-  let currentY = yOffset;
-  let rowMaxHeight = 0;
-
-  for (const panel of panelList) {
-    let dashboardPanel: DashboardPanel | null = null;
-
-    if (isLensAttachmentPanel(panel)) {
-      const config = panel.visualization as LensApiSchemaType;
-      const width = panel.grid?.w ?? getPanelWidth(config.type, normalizedLayout);
-      const height = panel.grid?.h ?? normalizedLayout.defaultPanelHeight;
-
-      if (currentX + width > normalizedLayout.dashboardGridColumnCount) {
-        currentX = 0;
-        currentY += rowMaxHeight;
-        rowMaxHeight = 0;
-      }
-      rowMaxHeight = Math.max(rowMaxHeight, height);
-
-      dashboardPanel = buildLensPanelFromApi(
-        config,
-        {
-          x: currentX,
-          y: currentY,
-          w: width,
-          h: height,
-        },
-        includePanelIdAsUid ? panel.panelId : undefined
-      );
-      currentX += width;
-    } else if (isGenericAttachmentPanel(panel)) {
-      dashboardPanel = buildPanelFromRawConfig({
-        embeddableType: panel.type,
-        rawConfig: panel.rawConfig,
-        title: panel.title,
-        position: {
-          currentX,
-          currentY,
-        },
-        layout: normalizedLayout,
-        uid: includePanelIdAsUid ? panel.panelId : undefined,
-        getLensPanelWidth,
-        preferredGrid: panel.grid,
-      });
-
-      if (dashboardPanel) {
-        currentX += dashboardPanel.grid.w;
-        rowMaxHeight = Math.max(rowMaxHeight, dashboardPanel.grid.h);
-        if (currentX >= normalizedLayout.dashboardGridColumnCount) {
-          currentX = 0;
-          currentY += rowMaxHeight;
-          rowMaxHeight = 0;
-        }
-      }
-    }
-
-    if (dashboardPanel) {
-      dashboardPanels.push(dashboardPanel);
-    }
-  }
-
-  return dashboardPanels;
-};
-
-export interface DashboardCanvasInitialInput {
+interface DashboardCanvasInitialInput {
   timeRange: {
     from: string;
     to: string;
@@ -269,7 +29,7 @@ export interface DashboardCanvasInitialInput {
   description?: string;
 }
 
-export const createDashboardRendererInitialInput = (
+const createDashboardRendererInitialInput = (
   data: DashboardAttachmentData
 ): DashboardCanvasInitialInput => ({
   timeRange: { from: 'now-24h', to: 'now' },
@@ -279,7 +39,7 @@ export const createDashboardRendererInitialInput = (
   description: data.description,
 });
 
-export const getDashboardRendererCreationOptions = async ({
+const getDashboardRendererCreationOptions = async ({
   savedObjectId,
   initialDashboardInput,
 }: {
@@ -310,11 +70,7 @@ export const saveDashboardFromApi = async (dashboardApi: DashboardApi): Promise<
   await dashboardApi.runInteractiveSave();
 };
 
-const getDashboardCanvasContentStyles = ({
-  euiTheme,
-}: {
-  euiTheme: ReturnType<typeof useEuiTheme>['euiTheme'];
-}) => ({
+const dashboardCanvasContentStyles = {
   root: css({
     display: 'flex',
     flexDirection: 'column',
@@ -327,24 +83,23 @@ const getDashboardCanvasContentStyles = ({
       display: 'none !important',
     },
   }),
-  actions: css({
-    padding: euiTheme.size.m,
-  }),
+  actions: ({ euiTheme }) =>
+    css({
+      padding: euiTheme.size.m,
+    }),
   renderer: css({
     flex: 1,
     minHeight: 0,
     display: 'flex',
   }),
-});
+};
 
 export const DashboardCanvasContent = ({
-  attachment,
+  attachment: { data },
 }: AttachmentRenderProps<DashboardAttachment>) => {
-  const { euiTheme } = useEuiTheme();
-  const data = attachment.data;
   const [dashboardApi, setDashboardApi] = useState<DashboardApi | undefined>();
   const [isSaveInProgress, setIsSaveInProgress] = useState(false);
-  const styles = useMemo(() => getDashboardCanvasContentStyles({ euiTheme }), [euiTheme]);
+  const styles = useMemoCss(dashboardCanvasContentStyles);
 
   const initialDashboardInput = useMemo(() => createDashboardRendererInitialInput(data), [data]);
